modified:   Game Workspace/game.py 	deleted:    Game Workspace/images/running.png 	deleted:    Game Workspace/images/sprite.png

+++ Game Workspace/game.py
@@ -19,7 +19,7 @@ class Enemy:
                 self.x += self.speed
             elif player_x < self.x:
                 self.x -= self.speed
-            else:
+            elif player_x == self.x: 
                 self.destroy
         else:
             self.x -= self.speed
